+++ packages/components/src/Form/Inputs/Combobox/utils/useComboboxRefs.ts
@@ -54,10 +54,12 @@ export function useComboboxRefs(forwardedRef: Ref<HTMLDivElement>) {
   const closeOnSelectPropRef = useRef(true)
   const windowedOptionsPropRef = useRef(false)
   const isAutoScrollingRef = useRef(false)
+  const indicatorPropRef = useRef(false)
 
   return {
     autoCompletePropRef,
     closeOnSelectPropRef,
+    indicatorPropRef,
     isAutoScrollingRef,
     listRef,
     optionsRef,