+++ packages/components/src/Modal/Prompt/Prompt.test.tsx
@@ -28,7 +28,7 @@ import { fireEvent, waitForElementToBeRemoved } from '@testing-library/react'
 import React, { useState } from 'react'
 
 import { renderWithTheme } from '@looker/components-test-utils'
-import { semanticColors, SemanticColors } from '@looker/design-tokens'
+import { theme, StatefulColor } from '@looker/design-tokens'
 import { Button } from '../../Button'
 import { Prompt } from './Prompt'
 
@@ -44,7 +44,7 @@ const requiredProps = {
 }
 
 const optionalProps = {
-  cancelColor: 'danger' as keyof SemanticColors,
+  cancelColor: 'critical' as StatefulColor,
   cancelLabel: 'Cancel Cheese',
   defaultValue: 'Default Value Cheese',
   onCancel: jest.fn(),
@@ -71,7 +71,7 @@ test('<Prompt/> with defaults', async () => {
 
   expect(input).toBeVisible()
   expect(getByText(requiredProps.title)).toBeVisible()
-  expect(saveButton).toHaveStyle(`background: ${semanticColors.primary.main}`)
+  expect(saveButton).toHaveStyle(`background: ${theme.colors.key}`)
 
   fireEvent.click(saveButton)
   expect(onSaveCallback).toHaveBeenCalledTimes(0)
@@ -99,9 +99,9 @@ test('<Prompt/> with custom props', () => {
   const cancelButton = getByText(optionalProps.cancelLabel)
 
   expect(cancelButton).toBeInTheDocument()
-  expect(cancelButton).toHaveStyle(`color: ${semanticColors.danger.main}`)
+  expect(cancelButton).toHaveStyle(`color: ${theme.colors.critical}`)
   expect(saveButton).toBeInTheDocument()
-  expect(saveButton).toHaveStyle(`background: ${semanticColors.primary.main}`)
+  expect(saveButton).toHaveStyle(`background: ${theme.colors.key}`)
   expect(getByDisplayValue(optionalProps.defaultValue)).toBeInTheDocument()
   expect(getByText('Secondary Cheese')).toBeInTheDocument()