+++ packages/design-tokens/src/tokens/color/colors.ts
@@ -24,77 +24,57 @@
 
  */
 
-import { SemanticColors } from '../../system'
+import { IntentColors, SurfaceColors } from '../../system/color/specifiable'
+import { CoreColors, Colors, SpecifiableColors } from '../../system'
 import { palette } from './palette'
+
+import { fallbackBlends, fallbackStateful } from './fallbacks'
 
 const {
   blue500,
-  charcoal000,
-  charcoal100,
-  charcoal300,
+  blue600,
   charcoal400,
-  charcoal500,
-  charcoal600,
-  primary500,
-  primary600,
-  primary700,
-  purple000,
-  purple100,
+  charcoal800,
+  green500,
   purple400,
-  purple500,
-  purple600,
-  red000,
-  red100,
-  red400,
   red500,
-  red600,
-  red700,
+  yellow500,
   white,
 } = palette
 
-export const semanticColors: SemanticColors = {
-  danger: {
-    altText: red400,
-    borderColor: charcoal300,
-    dark: red600,
-    darker: red700,
-    light: red100,
-    lighter: red000,
-    linkColor: blue500,
-    main: red500,
-    text: white,
-  },
-  neutral: {
-    altText: charcoal600,
-    borderColor: charcoal300,
-    dark: charcoal500,
-    darker: charcoal600,
-    light: charcoal100,
-    lighter: charcoal000,
-    linkColor: blue500,
-    main: charcoal400,
-    text: white,
-  },
-  primary: {
-    altText: purple400,
-    borderColor: charcoal300,
-    dark: purple500,
-    darker: purple600,
-    light: purple100,
-    lighter: purple000,
-    linkColor: blue500,
-    main: purple400,
-    text: white,
-  },
-  secondary: {
-    altText: purple400,
-    borderColor: charcoal300,
-    dark: primary600,
-    darker: primary700,
-    light: purple100,
-    lighter: purple000,
-    linkColor: blue500,
-    main: primary500,
-    text: white,
-  },
+/* eslint-disable sort-keys-fix/sort-keys-fix */
+
+const defaultCoreColors: CoreColors = {
+  background: white,
+  text: charcoal800,
+  key: purple400,
+}
+
+const generateSurfaceColors = (coreColors: CoreColors): SurfaceColors => {
+  return {
+    field: coreColors.background,
+    inverse: coreColors.text,
+    inverseText: coreColors.background,
+  }
+}
+
+const defaultIntentColors: IntentColors = {
+  link: blue600,
+  critical: red500,
+  warn: yellow500,
+  neutral: charcoal400,
+  positive: green500,
+  inform: blue500,
+}
+
+const specifiableColors: SpecifiableColors = {
+  ...defaultCoreColors,
+  ...generateSurfaceColors(defaultCoreColors),
+  ...defaultIntentColors,
+}
+
+export const colors: Colors = {
+  ...specifiableColors,
+  ...fallbackBlends,
+  ...fallbackStateful,
 }