+++ packages/test-utils/src/helpers/react-testing-library/index.ts
@@ -24,4 +24,10 @@
 
  */
 
-export * from './Banner'
+export {
+  closeCombobox,
+  getAllComboboxOptionText,
+  getComboboxOptions,
+  getComboboxOptionText,
+  openCombobox,
+} from './comboboxHelpers'